requirements.txt
fastapi==0.119.0
sqlalchemy==2.0.44
asyncpg==0.30.0
httpx==0.28.1
python-dotenv==1.1.1
uvicorn==0.38.0
passlib==1.7.4
pydantic==2.12.3
